Fixed bugs
Fixed bugs including:

CTRL + D was permenantly CTRL on once pressed once.

The preview can be buggy IF when sending in a new prop array there is an element in array1 in index K which matches array2 index K's type. I think it's something in the way that Javascript overwrites arrays